+++ src/main/resources/templates/fragments/header.html
@@ -12,39 +12,36 @@
                 </li>
             </ul>
             <ul class="nav-right">
-                <li id="header-user-dropdown-btn" class="user-profile dropdown no-border pointer">
-<!--                <li th:if="${user}" id="header-user-dropdown-btn" class="user-profile dropdown no-border pointer">-->
+                <li th:if="${user}" id="header-user-dropdown-btn" class="user-profile dropdown no-border pointer">
                     <a class="user-info-link user-dropdown-menu dropdown-toggle" data-toggle="dropdown">
-                        <img class="profile-img img-fluid" src="https://avatars3.githubusercontent.com/u/46308949?s=460&v=4" th:alt="brown">
-<!--                        <img class="profile-img img-fluid" src="https://avatars3.githubusercontent.com/u/46308949?s=460&v=4" th:alt="${user.userName}">-->
+                        <img class="profile-img img-fluid" src="https://avatars3.githubusercontent.com/u/46308949?s=460&v=4" th:alt="${user.name}">
                         <div class="user-info text-white">
-                            <span class="name pdd-right-5">brown</span>
-<!--                            <span class="name pdd-right-5" th:text="${user.userName}"></span>-->
+                            <span class="name pdd-right-5" th:text="${user.name}"></span>
                         </div>
                     </a>
                     <ul class="dropdown-menu">
                         <li class="text-dark">
-                            <a href="/mypage.html">
+                            <a href="/mypage">
                                 <i class="ti-user pdd-right-10"></i>
                                 <span class="text-dark">마이 페이지</span>
                             </a>
                         </li>
                         <li role="separator" class="divider"></li>
                         <li class="text-dark">
-                            <a href="/">
+                            <a href="/logout">
                                 <i class="ti-power-off pdd-right-10"></i>
                                 <span class="text-dark">로그아웃</span>
                             </a>
                         </li>
                     </ul>
                 </li>
-<!--                <li th:unless="${user}" class="user-profile">-->
-<!--                    <a href="/login">-->
-<!--                        <div class="user-info login-btn pointer mrg-right-10">-->
-<!--                            <span class="name pdd-right-5 text-white">Login</span>-->
-<!--                        </div>-->
-<!--                    </a>-->
-<!--                </li>-->
+                <li th:unless="${user}" class="user-profile">
+                    <a href="/login">
+                        <div class="user-info login-btn pointer mrg-right-10">
+                            <span class="name pdd-right-5 text-white">Login</span>
+                        </div>
+                    </a>
+                </li>
             </ul>
         </div>
     </header>
